build: add pkg-config support
With this patch CMake will now generate a pkg-config file that users
can use to discover and consume the zarchive library.

pkg-config was chosen instead of a CMake Config file as it is available
on more platform and is build system agnostic, meaning that you can use
it also if you're not using CMake.

To see why JoinPaths is required, see
https://github.com/jtojnar/cmake-snips#concatenating-paths-when-building-pkg-config-files